requirements.txt
eth-brownie>=1.13.4
web3>=5.11.1
click>=7.1.2
tabulate>=0.8.7
joblib>=1.0.1
cachetools>=4.1.1
fastapi>=0.63.0
uvicorn>=0.13.4
pony>=0.7.13
psycopg2-binary>=2.8.5
tokenlists>=0.1.0b0
aiofiles>=0.6.0